P for roll and pitch, motor bug

+++ software/lib/ESC/esc.h
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <math.h>
 #include "pico/stdlib.h"
 #include "hardware/irq.h"
 #include "hardware/clocks.h"
@@ -13,12 +14,12 @@ typedef struct ESC
 
 } ESC;
 
-void esc_setup(ESC *myesc, uint32_t *PIN_PWM, uint32_t PWM_FREQ, uint32_t PWM_WRAP, double MIN_DUTY_CYC, double MAX_DUTY_CYC);
+void esc_setup(ESC *myesc, uint32_t *PIN_PWM, uint32_t PWM_FREQ, uint32_t PWM_WRAP, float MIN_DUTY_CYC, float MAX_DUTY_CYC);
 
-//calibrate the ESC so it know the min and max throttle
+// calibrate the ESC so it know the min and max throttle
 void cali_motor(ESC *myesc);
 // arm the ESC so it can starts
 void arm_motor(ESC *myesc);
 
 // percent_throttle: [0,1], 0 for no throttle, 1 for max throttle
-void motor_control(ESC *myesc, double percent_throttle, uint MOTOR_NUM);
+void motor_control(ESC *myesc, float percent_throttle, uint MOTOR_NUM);
